requirements.txt
crawl4ai>=0.3.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
